# eliminate uninferrable `typeof(x) === T` conditions