Most of these conditions were introduced in #25828 and #30480 for some
performance reasons atm, but now they seem just unnecessary or even
harmful in terms of inferrability.

There doesn't seem to be any performance difference in the benchmark
used at #25828:
```julia
using BenchmarkTools
x = rand(Int, 100_000);
y = convert(Vector{Union{Int,Missing}}, x);
z = copy(y); z[2] = missing;
```

> master:
```julia
julia> @btime map(identity, x);
  57.814 μs (3 allocations: 781.31 KiB)

julia> @btime map(identity, y);
  94.040 μs (3 allocations: 781.31 KiB)

julia> @btime map(identity, z);
  127.554 μs (5 allocations: 1.62 MiB)

julia> @btime broadcast(x->x, x);
  59.248 μs (2 allocations: 781.30 KiB)

julia> @btime broadcast(x->x, y);
  74.693 μs (2 allocations: 781.30 KiB)

julia> @btime broadcast(x->x, z);
  126.262 μs (4 allocations: 1.62 MiB)
```

> this commit:
```
julia> @btime map(identity, x);
  58.668 μs (3 allocations: 781.31 KiB)

julia> @btime map(identity, y);
  94.013 μs (3 allocations: 781.31 KiB)

julia> @btime map(identity, z);
  126.600 μs (5 allocations: 1.62 MiB)

julia> @btime broadcast(x->x, x);
  57.531 μs (2 allocations: 781.30 KiB)

julia> @btime broadcast(x->x, y);
  69.561 μs (2 allocations: 781.30 KiB)

julia> @btime broadcast(x->x, z);
  125.578 μs (4 allocations: 1.62 MiB)
```

---

@nanosoldier `runbenchmarks("union", vs=":master")`