@nanosoldier 2022-04-18
[Your benchmark job](https://github.com/JuliaLang/julia/pull/45018) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/6ec1ef4_vs_1600cb9/report.md).