@aviatesk 2022-04-18
> Can you also check when there are e.g. 25-50% of missing values just in case?

Confirmed there seems to be no difference.